requirements.txt
asgiref==3.9.1
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2
djangorestframework==3.14.0
django-cors-headers==3.13.0

djangorestframework_simplejwt==5.2.0
django-axes==5.25.0